Get rid of internal/clock #1 - Consolidate AfterFunc and Sleep in tests
Resolves #3

Was looking at the open issues and #3 seemed to make sense - it was
also something that I saw in #37 so I thought I'd take a stab.

The reasoning is that less code + smaller interface is a good thing.

During the diff it turned out that:
- our internal/clock has no tests
- our internal/clock has at least two bugs:
-- one that I'm fixing here, related to not executing two timers
   if they're scheduled at the "now" boundry
-- one regarding adding time but not having a timer set for that
   specific time (not a problem for us, so not touching it)

Looking at the internal/clock, we see:
```
// Forked from github.com/andres-erbsen/clock to isolate a missing nap.
```

There's one extra nap in AfterFunc that I'm removing in this diff,
and so I'm hoping that by removing AfterFunc we'll eventually be able
to move off the fork.

The tests are a bit of a pain, I'll put up another PR for a test
runner that will make this a bit easier.